Basic Input and Output Program
This program requires the student to create a short and program that asks the user for three numbers.  The output will be the sum and average of the three numbers.
Sample output (you must follow this example):
The three numbers are: 99, 99, and 99.
The sum of the numbers is: 99.
The average of the numbers is: 99.99.