10/29/2018 Methods and File Output and Exceptions
Program Specifications:
You to write a menu driven program.  The program will use a switch statement.  The cases will be as follows:
Get the user’s first name and echo it back out 20 times.
Get the Store user’s age and double it and display the age and the doubled age.
 Using the age from #2 output one of the following statements
Since you are XX years old, you are a teenager
Since you are YY years old, you are NOT a teenager
Get a single integer between 3 and 50 (inclusively) from the user.  Create a triangle of X’s with the integer inputted rows.  The triangle needs to be displayed on the screen and in a text document named triangle.txt.